requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==3.1.0
defusedxml==0.7.1
Django==3.2
django-cors-headers==4.3.0
django-rest-authtoken==2.1.4
djangorestframework==3.14.0
et-xmlfile==1.1.0
httpie==3.2.1
idna==3.4
markdown-it-py==2.2.0
mdurl==0.1.2
multidict==6.0.4
numpy==1.23.2
oauthlib==3.2.2
openpyxl==3.0.10
pandas==1.4.4
Pygments==2.15.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.29.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.3.5
six==1.16.0
sqlparse==0.4.2
tweepy==4.14.0
urllib3==1.26.15
whitenoise==6.4.0
xlrd==2.0.1
